Hide UCrop behind abstraction to decouple AvatarPicker
